Implement TestHarness
TestHarness is a class which runs multiple tests and collects statistics
about them. It is inspired by perl's Test::Harness. It is the main
component of a program which tests other programs which output TAP.

The current design I'm implementing requires that unit tests are
implemented in classes outside of the unit they are testing, for the
following reasons:

  1. You should only test public interface anyway. Even if it is
     sometimes simpler to implement a test using private data of
     an object:
  2. It makes the class unnecessarily big and exposes methods which
     you don't want/need for a release.

This base class for implementing tests comes next. Tests are run in
an external process, by using the test as a startup class. This has
the benefit that the test harness isn't affected by a crash in the
test. It would be nice to have special treatment of tests in the
project directory structure, to list all tests in a project reliably
from the outside. IDE and compiler could benefit from that as well.